Close #9: Let user get rid of dialog without opting-in
It is important to allow users to get rid of the cookie opt-in dialog;
otherwise they easily may feel coerced to accept something they
actually don't want.  Therefore, we change the privacy form to have two
buttons, one to accept cookies, and another to decline them.  If the
accept button is clicked, it works as before when the checkbox was
activated (we don't consider the additional checkbox to be necessary,
or even sensible anymore; pressing a button is sufficient to opt-in);
if the decline button is clicked, we also set the `privacy_agreed`
cookie, but in this case with the value `no` and an expiration time of
`0`.  This means the cookie is a session cookie, which is deleted when
the browser is closed.  We consider this cookie to be a user-input
cookie which are exempt from cookie consent, anyway.

Since the value of the `privacy_agreed` cookie is now relevant, it has
to be checked for privacy guards in the template, which is a BC break.
Therefore, we abstract (which we should have done in the first place),
and introduce the function `privacy_agreed()` which should be used
instead of the `isset($_COOKIE['privacy_agreed'])`.